+++ vm/asyncio.go
@@ -155,6 +155,8 @@ func (i *Interp) builtinModule(name string) (*object.Module, bool) {
 		return i.buildLocale(), true
 	case "pydoc":
 		return i.buildPydoc(), true
+	case "faulthandler":
+		return i.buildFaulthandler(), true
 	case "xmlrpc":
 		return i.buildXmlrpc(), true
 	case "xmlrpc.client":